Moved custom exceptions back to models directory
