+++ src/umleditor/mvc_controller/controller_input.py
@@ -1,4 +1,4 @@
-from umleditor.custom_exceptions CustomExceptions as CE
+from umleditor.mvc_model.custom_exceptions CustomExceptions as CE
 from umleditor.mvc_controller.autofill import CommandCompleter
 from prompt_toolkit import prompt
 from prompt_toolkit.history import InMemoryHistory
