+++ src/umleditor/mvc_controller/momento.py
@@ -1,5 +1,5 @@
 import os
-from umleditor.custom_exceptions import CustomExceptions as CE
+from umleditor.mvc_model.custom_exceptions import CustomExceptions as CE
 
 
 class Momento:
