+++ src/umleditor/mvc_model/entity.py
@@ -1,7 +1,7 @@
 # Primary: Danish
 # Secondary: Zhang
 
-from umleditor.custom_exceptions import CustomExceptions
+from umleditor.mvc_model.custom_exceptions import CustomExceptions
 
 
 class Entity:
